refactor: Separate TestMain component
- Modify test routes
